caches

+++ src/LRUCacheSimple.cpp
@@ -16,11 +16,12 @@ int LRUCache::get(int key) {
 void LRUCache::put(int key, int value) {
   if (capacity == 0)
     return;
-  auto it = cache.find(key);
+  auto it = cache.find(key); // to check if the key already exists and update it
 
   if (it != cache.end()) {
-    order.erase(it->second.second);
-  } else if (cache.size() == capacity) {
+    order.erase(it->second.second); // if the key is arlreay there erase it & we
+                                    // will udpate in below steps
+  } else if (cache.size() == static_cast<size_t>(capacity)) {
     int lru = order.front();
     order.pop_front();
     cache.erase(lru);
